refactor(pipeline-print): move the pipeline print function to the correct file
