+++ src/cli/repl/server/connection.ts
@@ -24,9 +24,8 @@ import { cfg2quads, ControlFlowInformation, extractCFG } from '../../../util/cfg
 import { defaultQuadIdGenerator, QuadSerializationConfiguration } from '../../../util/quads'
 import { deepMergeObject } from '../../../util/objects'
 import { LogLevel } from '../../../util/log'
-import { StepOutputFormat } from '../../../core/print/print'
+import { printStepResult, StepOutputFormat } from '../../../core/print/print'
 import { DataflowInformation } from '../../../dataflow/internal/info'
-import { printStepResult } from '../../../core/steps/print'
 import { PARSE_WITH_R_SHELL_STEP } from '../../../core/steps/all/00-parse'
 import { NORMALIZE } from '../../../core/steps/all/10-normalize'
 import { LEGACY_STATIC_DATAFLOW } from '../../../core/steps/all/20-dataflow'
