+++ src/cli/statistics-helper-app.ts
@@ -12,9 +12,8 @@ import { create } from 'tar'
 import fs from 'fs'
 import { guard } from '../util/assert'
 import { retrieveArchiveName } from './common/features'
-import { StepOutputFormat } from '../core/print/print'
+import { printStepResult, StepOutputFormat } from '../core/print/print'
 import { date2string } from '../util/time'
-import { printStepResult } from '../core/steps/print'
 import { PARSE_WITH_R_SHELL_STEP } from '../core/steps/all/00-parse'
 import { NORMALIZE } from '../core/steps/all/10-normalize'
 import { LEGACY_STATIC_DATAFLOW } from '../core/steps/all/20-dataflow'
